requirements.txt
django>=4.2
djangorestframework>=3.14
python-dotenv>=1.0
google-generativeai>=0.5
